[blockchain] refs #102 - Move sign service into a different package
