+++ exchange_app/api_1_0/__init__.py
@@ -2,4 +2,4 @@ from flask import Blueprint
 
 api = Blueprint('api', __name__)
 
-from . import address, wallets, assets, pending_events, capabilities, isalive, common, balances, history, explorer_url, signservice, transactions
+from . import address, wallets, assets, pending_events, capabilities, isalive, balances, history, explorer_url, transactions
