+++ exchange_app/common.py
