+++ exchange_app/signservice/__init__.py
@@ -1,0 +1,3 @@
+from flask import Blueprint
+
+from . import signservice
